CLN: rename package to force_gromacs
