+++ force_gromacs/commands/tests/test_gromacs_commands.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 from traits.trait_errors import TraitError
 
-from gromacs.api import (
+from force_gromacs.api import (
     Gromacs_genbox, Gromacs_grompp, Gromacs_genion,
     Gromacs_mdrun, Gromacs_genconf
 )
