+++ force_gromacs/core/tests/test_base_gromacs_command.py
@@ -2,7 +2,7 @@ from unittest import TestCase
 
 from traits.trait_errors import TraitError
 
-from gromacs.core.base_gromacs_command import (
+from force_gromacs.core.base_gromacs_command import (
     BaseGromacsCommand
 )
