+++ force_gromacs/data_sources/chemical/chemical_model.py
